oss/requirements.txt
annotated-types==0.6.0
autocommand==2.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
cheroot==10.0.0
CherryPy==18.9.0
cherrypy-cors==1.7.0
dnspython==2.6.1
enum34==1.1.10
httpagentparser==1.9.5
idna==3.6
inflect==7.0.0
jaraco.collections==5.0.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
kafka-python==2.0.2
lxml==5.1.0
marshmallow==3.20.2
more-itertools==10.1.0
packaging==23.2
pika==1.3.2
portend==3.2.0
pyaml==23.12.0
pyang==2.6.0
pyangbind==0.8.5
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
repoze.lru==0.7
requests==2.31.0
Routes==2.5.1
six==1.16.0
tempora==5.5.0
typing_extensions==4.9.0
urllib3==2.1.0
zc.lockfile==3.0.post1
